handle mailto addresses in WebViewControllers
